requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.37.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.44
alembic==1.17.1
psycopg2-binary==2.9.11

# Auth
auth0-python==4.13.0
python-jose[cryptography]==3.5.0

# AI
anthropic==0.72.0
openai==1.109.1

# Email
sendgrid==6.12.5
python-decouple==3.8


# Utilities
pydantic==2.12.3
pydantic-settings==2.11.0
python-dotenv==1.1.1
httpx==0.28.1
# Optional (add later)
# redis==5.0.1
# celery==5.3.6